+++ src/http_echo.rs
@@ -3,7 +3,7 @@ use std::io::{BufRead, BufReader, Write};
 use std::net::{SocketAddr, TcpListener, TcpStream};
 use std::thread;
 
-pub fn http_echo(addr: &SocketAddr) -> Result<(), std::io::Error> {
+pub fn http_echo(addr: &SocketAddr, liveness_probe_path: String) -> Result<(), std::io::Error> {
     let listener = TcpListener::bind(addr)?;
     println!("Listening for HTTP connections on {addr}");
 
@@ -15,7 +15,8 @@ pub fn http_echo(addr: &SocketAddr) -> Result<(), std::io::Error> {
                         "Accepting HTTP connection from {}",
                         stream.peer_addr().expect("Unable to obtain peer address")
                     );
-                    thread::spawn(move || handle_client(stream));
+                    let liveness_probe_path = liveness_probe_path.clone();
+                    thread::spawn(move || handle_client(stream, liveness_probe_path));
                 }
                 Err(e) => {
                     eprintln!("Error accepting HTTP connection: {e}");
@@ -26,7 +27,7 @@ pub fn http_echo(addr: &SocketAddr) -> Result<(), std::io::Error> {
     Ok(())
 }
 
-fn handle_client(stream: TcpStream) {
+fn handle_client(stream: TcpStream, liveness_probe_path: String) {
     let mut reader = BufReader::new(stream);
 
     loop {
@@ -40,6 +41,9 @@ fn handle_client(stream: TcpStream) {
         if request_line == "\r\n" || request_line == "\n" {
             continue;
         }
+
+        // Extract the path from the request line.
+        let path = request_line.split_whitespace().nth(1).unwrap_or("/");
 
         // Read and ignore headers.
         let mut header_line = String::new();
@@ -63,8 +67,14 @@ fn handle_client(stream: TcpStream) {
         let version = env!("CARGO_PKG_VERSION");
         let date = Utc::now().format("%a, %d %b %Y %T GMT").to_string();
 
-        let response = format!(
-            "HTTP/1.1 204 No Content\r\nServer: crashie/{version}\r\nDate: {date}\r\nContent-Length: 0\r\nCache-Control: no-cache, no-store\r\n\r\n");
+        // Prepare response based on the request path
+        let response = if path == liveness_probe_path {
+            format!(
+                "HTTP/1.1 200 OK\r\nServer: crashie/{version}\r\nDate: {date}\r\nContent-Length: 0\r\nCache-Control: no-cache, no-store\r\n\r\n")
+        } else {
+            format!(
+                "HTTP/1.1 204 No Content\r\nServer: crashie/{version}\r\nDate: {date}\r\nContent-Length: 0\r\nCache-Control: no-cache, no-store\r\n\r\n")
+        };
 
         if let Err(e) = stream.write_all(response.as_bytes()) {
             eprintln!("Failed to write HTTP response: {e}")